+++ android/src/main/java/com/brentvatne/exoplayer/ReactExoplayerView.java
@@ -1,5 +1,10 @@
 package com.brentvatne.exoplayer;
 
+import static com.google.android.exoplayer2.C.CONTENT_TYPE_DASH;
+import static com.google.android.exoplayer2.C.CONTENT_TYPE_HLS;
+import static com.google.android.exoplayer2.C.CONTENT_TYPE_OTHER;
+import static com.google.android.exoplayer2.C.CONTENT_TYPE_SS;
+
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.ActivityManager;
@@ -17,26 +22,23 @@ import android.view.accessibility.CaptioningManager;
 import android.widget.FrameLayout;
 import android.widget.ImageButton;
 
+import androidx.annotation.WorkerThread;
 import androidx.activity.OnBackPressedCallback;
 
+import com.brentvatne.common.Track;
+import com.brentvatne.common.VideoTrack;
 import com.brentvatne.react.R;
 import com.brentvatne.receiver.AudioBecomingNoisyReceiver;
 import com.brentvatne.receiver.BecomingNoisyListener;
-import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Dynamic;
 import com.facebook.react.bridge.LifecycleEventListener;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
-import com.facebook.react.bridge.WritableArray;
-import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.uimanager.ThemedReactContext;
-import com.facebook.react.util.RNLog;
+import com.google.ads.interactivemedia.v3.api.AdEvent;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.DefaultLoadControl;
 import com.google.android.exoplayer2.DefaultRenderersFactory;
-import com.google.android.exoplayer2.ExoPlaybackException;
-import com.google.android.exoplayer2.drm.MediaDrmCallbackException;
-import com.google.android.exoplayer2.drm.DrmSession.DrmSessionException;
 import com.google.android.exoplayer2.Format;
 import com.google.android.exoplayer2.MediaItem;
 import com.google.android.exoplayer2.PlaybackException;
@@ -50,16 +52,12 @@ import com.google.android.exoplayer2.drm.DefaultDrmSessionManagerProvider;
 import com.google.android.exoplayer2.drm.DrmSessionEventListener;
 import com.google.android.exoplayer2.drm.DrmSessionManager;
 import com.google.android.exoplayer2.drm.DrmSessionManagerProvider;
-import com.google.android.exoplayer2.drm.ExoMediaDrm;
 import com.google.android.exoplayer2.drm.FrameworkMediaDrm;
 import com.google.android.exoplayer2.drm.HttpMediaDrmCallback;
-import com.google.android.exoplayer2.drm.MediaDrmCallbackException;
 import com.google.android.exoplayer2.drm.UnsupportedDrmException;
 import com.google.android.exoplayer2.mediacodec.MediaCodecInfo;
-import com.google.android.exoplayer2.mediacodec.MediaCodecRenderer;
 import com.google.android.exoplayer2.mediacodec.MediaCodecUtil;
 import com.google.android.exoplayer2.metadata.Metadata;
-import com.google.android.exoplayer2.source.BehindLiveWindowException;
 import com.google.android.exoplayer2.source.MediaSource;
 import com.google.android.exoplayer2.source.MergingMediaSource;
 import com.google.android.exoplayer2.source.ProgressiveMediaSource;
@@ -75,15 +73,15 @@ import com.google.android.exoplayer2.trackselection.AdaptiveTrackSelection;
 import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
 import com.google.android.exoplayer2.trackselection.MappingTrackSelector;
 import com.google.android.exoplayer2.trackselection.ExoTrackSelection;
-import com.google.android.exoplayer2.trackselection.TrackSelectionParameters;
+import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.trackselection.TrackSelectionOverride;
 import com.google.android.exoplayer2.ui.PlayerControlView;
 import com.google.android.exoplayer2.upstream.BandwidthMeter;
 import com.google.android.exoplayer2.upstream.DataSource;
+import com.google.android.exoplayer2.upstream.DataSpec;
 import com.google.android.exoplayer2.upstream.DefaultAllocator;
 import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
 import com.google.android.exoplayer2.upstream.HttpDataSource;
-import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Util;
 import com.google.android.exoplayer2.trackselection.TrackSelection;
 import com.google.android.exoplayer2.source.dash.DashUtil;
@@ -91,8 +89,12 @@ import com.google.android.exoplayer2.source.dash.manifest.DashManifest;
 import com.google.android.exoplayer2.source.dash.manifest.Period;
 import com.google.android.exoplayer2.source.dash.manifest.AdaptationSet;
 import com.google.android.exoplayer2.source.dash.manifest.Representation;
-import com.google.android.exoplayer2.source.dash.manifest.Descriptor;
 
+import com.google.android.exoplayer2.ext.ima.ImaAdsLoader;
+import com.google.android.exoplayer2.source.ads.AdsMediaSource;
+import com.google.android.exoplayer2.source.DefaultMediaSourceFactory;
+
+import com.google.common.collect.ImmutableList;
 import java.net.CookieHandler;
 import java.net.CookieManager;
 import java.net.CookiePolicy;
@@ -101,9 +103,6 @@ import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 import java.util.Map;
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.List;
 import java.lang.Thread;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -119,7 +118,8 @@ class ReactExoplayerView extends FrameLayout implements
         BandwidthMeter.EventListener,
         BecomingNoisyListener,
         AudioManager.OnAudioFocusChangeListener,
-        DrmSessionEventListener {
+        DrmSessionEventListener,
+        AdEvent.AdEventListener {
 
     public static final double DEFAULT_MAX_HEAP_ALLOCATION_PERCENT = 1;
     public static final double DEFAULT_MIN_BACK_BUFFER_MEMORY_RESERVE = 0;
@@ -144,6 +144,7 @@ class ReactExoplayerView extends FrameLayout implements
 
     private ExoPlayerView exoPlayerView;
     private FullScreenPlayerView fullScreenPlayerView;
+    private ImaAdsLoader adsLoader;
 
     private DataSource.Factory mediaDataSourceFactory;
     private ExoPlayer player;
@@ -203,6 +204,7 @@ class ReactExoplayerView extends FrameLayout implements
     private String drmLicenseUrl = null;
     private String[] drmLicenseHeader = null;
     private boolean controls;
+    private Uri adTagUrl;
     // \ End props
 
     // React
@@ -221,6 +223,9 @@ class ReactExoplayerView extends FrameLayout implements
             switch (msg.what) {
                 case SHOW_PROGRESS:
                     if (player != null) {
+                        if (playerControlView != null && isPlayingAd() && controls) {
+                            playerControlView.hide();
+                        }
                         long pos = player.getCurrentPosition();
                         long bufferedDuration = player.getBufferedPercentage() * player.getDuration() / 100;
                         long duration = player.getDuration();
@@ -263,6 +268,9 @@ class ReactExoplayerView extends FrameLayout implements
         audioBecomingNoisyReceiver = new AudioBecomingNoisyReceiver(themedReactContext);
     }
 
+    private boolean isPlayingAd() {
+        return player != null && player.isPlayingAd();
+    }
 
     @Override
     public void setId(int id) {
@@ -372,6 +380,15 @@ class ReactExoplayerView extends FrameLayout implements
             playerControlView = new PlayerControlView(getContext());
         }
 
+        if (fullScreenPlayerView == null) {
+            fullScreenPlayerView = new FullScreenPlayerView(getContext(), exoPlayerView, playerControlView, new OnBackPressedCallback(true) {
+                @Override
+                public void handleOnBackPressed() {
+                    setFullscreen(false);
+                }
+            });
+        }
+
         // Setting the player for the playerControlView
         playerControlView.setPlayer(player);
         playPauseControlContainer = playerControlView.findViewById(R.id.exo_play_pause_container);
@@ -380,7 +397,9 @@ class ReactExoplayerView extends FrameLayout implements
         exoPlayerView.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View v) {
-                togglePlayerControlVisibility();
+                if (!isPlayingAd()) {
+                    togglePlayerControlVisibility();
+                }
             }
         });
 
@@ -406,8 +425,9 @@ class ReactExoplayerView extends FrameLayout implements
         });
 
         //Handling the fullScreenButton click event
-        ImageButton fullScreenButton = playerControlView.findViewById(R.id.exo_fullscreen);
+        final ImageButton fullScreenButton = playerControlView.findViewById(R.id.exo_fullscreen);
         fullScreenButton.setOnClickListener(v -> setFullscreen(!isFullscreen));
+        updateFullScreenButtonVisbility();
 
         // Invoking onPlaybackStateChanged and onPlayWhenReadyChanged events for Player
         eventListener = new Player.Listener() {
@@ -421,7 +441,6 @@ class ReactExoplayerView extends FrameLayout implements
                 if (pauseButton != null && pauseButton.getVisibility() == GONE) {
                     pauseButton.setVisibility(INVISIBLE);
                 }
-
                 reLayout(playPauseControlContainer);
                 //Remove this eventListener once its executed. since UI will work fine once after the reLayout is done
                 player.removeListener(eventListener);
@@ -441,7 +460,7 @@ class ReactExoplayerView extends FrameLayout implements
      * Adding Player control to the frame layout
      */
     private void addPlayerControl() {
-        if(player == null) return;
+        if(playerControlView == null) return;
         LayoutParams layoutParams = new LayoutParams(
                 LayoutParams.MATCH_PARENT,
                 LayoutParams.MATCH_PARENT);
@@ -546,6 +565,12 @@ class ReactExoplayerView extends FrameLayout implements
                                     eventEmitter.error("Failed to initialize DRM Session Manager Framework!", new Exception("DRM Session Manager Framework failure!"), "3003");
                                     return;
                                 }
+                                    
+                                if (activity == null) {
+                                    Log.e("ExoPlayer Exception", "Failed to initialize Player!");
+                                    eventEmitter.error("Failed to initialize Player!", new Exception("Current Activity is null!"), "1001");
+                                    return;
+                                }
 
                                 // Initialize handler to run on the main thread
                                 activity.runOnUiThread(new Runnable() {
@@ -598,13 +623,24 @@ class ReactExoplayerView extends FrameLayout implements
         DefaultRenderersFactory renderersFactory =
                 new DefaultRenderersFactory(getContext())
                         .setExtensionRendererMode(DefaultRenderersFactory.EXTENSION_RENDERER_MODE_OFF);
+
+        // Create an AdsLoader.
+        adsLoader = new ImaAdsLoader.Builder(themedReactContext).setAdEventListener(this).build();
+
+        MediaSource.Factory mediaSourceFactory = new DefaultMediaSourceFactory(mediaDataSourceFactory)
+            .setLocalAdInsertionComponents(unusedAdTagUri -> adsLoader, exoPlayerView);
+
         player = new ExoPlayer.Builder(getContext(), renderersFactory)
                     .setTrackSelector(self.trackSelector)
                     .setBandwidthMeter(bandwidthMeter)
                     .setLoadControl(loadControl)
+                    .setMediaSourceFactory(mediaSourceFactory)
                     .build();
         player.addListener(self);
         exoPlayerView.setPlayer(player);
+        if (adsLoader != null) {
+            adsLoader.setPlayer(player);
+        }
         audioBecomingNoisyReceiver.setListener(self);
         bandwidthMeter.addEventListener(new Handler(), self);
         setPlayWhenReady(!isPaused);
@@ -634,11 +670,26 @@ class ReactExoplayerView extends FrameLayout implements
     private void initializePlayerSource(ReactExoplayerView self, DrmSessionManager drmSessionManager) {
         ArrayList<MediaSource> mediaSourceList = buildTextSources();
         MediaSource videoSource = buildMediaSource(self.srcUri, self.extension, drmSessionManager);
+        MediaSource mediaSourceWithAds = null;
+        if (adTagUrl != null) {
+            MediaSource.Factory mediaSourceFactory = new DefaultMediaSourceFactory(mediaDataSourceFactory)
+                    .setLocalAdInsertionComponents(unusedAdTagUri -> adsLoader, exoPlayerView);
+            DataSpec adTagDataSpec = new DataSpec(adTagUrl);
+            mediaSourceWithAds = new AdsMediaSource(videoSource, adTagDataSpec, ImmutableList.of(srcUri, adTagUrl), mediaSourceFactory, adsLoader, exoPlayerView);
+        }
         MediaSource mediaSource;
         if (mediaSourceList.size() == 0) {
-            mediaSource = videoSource;
+            if (mediaSourceWithAds != null) {
+                mediaSource = mediaSourceWithAds;
+            } else {
+                mediaSource = videoSource;
+            }
         } else {
-            mediaSourceList.add(0, videoSource);
+            if (mediaSourceWithAds != null) {
+                mediaSourceList.add(0, mediaSourceWithAds);
+            } else {
+                mediaSourceList.add(0, videoSource);
+            }
             MediaSource[] textSourceArray = mediaSourceList.toArray(
                     new MediaSource[mediaSourceList.size()]
             );
@@ -675,12 +726,6 @@ class ReactExoplayerView extends FrameLayout implements
         setControls(controls);
         applyModifiers();
         startBufferCheckTimer();
-        fullScreenPlayerView = new FullScreenPlayerView(getContext(), exoPlayerView, playerControlView, new OnBackPressedCallback(true) {
-            @Override
-            public void handleOnBackPressed() {
-                setFullscreen(false);
-            }
-        });
     }
 
     private DrmSessionManager buildDrmSessionManager(UUID uuid, String licenseUrl, String[] keyRequestPropertiesArray) throws UnsupportedDrmException {
@@ -726,7 +771,17 @@ class ReactExoplayerView extends FrameLayout implements
         int type = Util.inferContentType(!TextUtils.isEmpty(overrideExtension) ? "." + overrideExtension
                 : uri.getLastPathSegment());
         config.setDisableDisconnectError(this.disableDisconnectError);
-        MediaItem mediaItem = new MediaItem.Builder().setUri(uri).build();
+
+        MediaItem.Builder mediaItemBuilder = new MediaItem.Builder().setUri(uri);
+
+        if (adTagUrl != null) {
+            mediaItemBuilder.setAdsConfiguration(
+                    new MediaItem.AdsConfiguration.Builder(adTagUrl).build()
+            );
+        }
+
+        MediaItem mediaItem = mediaItemBuilder.build();
+
         DrmSessionManagerProvider drmProvider = null;
         if (drmSessionManager != null) {
             drmProvider = new DrmSessionManagerProvider() {
@@ -739,7 +794,7 @@ class ReactExoplayerView extends FrameLayout implements
             drmProvider = new DefaultDrmSessionManagerProvider();
         }
         switch (type) {
-            case C.TYPE_SS:
+            case CONTENT_TYPE_SS:
                 return new SsMediaSource.Factory(
                         new DefaultSsChunkSource.Factory(mediaDataSourceFactory),
                         buildDataSourceFactory(false)
@@ -747,7 +802,7 @@ class ReactExoplayerView extends FrameLayout implements
                  .setLoadErrorHandlingPolicy(
                         config.buildLoadErrorHandlingPolicy(minLoadRetryCount)
                 ).createMediaSource(mediaItem);
-            case C.TYPE_DASH:
+            case CONTENT_TYPE_DASH:
                 return new DashMediaSource.Factory(
                         new DefaultDashChunkSource.Factory(mediaDataSourceFactory),
                         buildDataSourceFactory(false)
@@ -755,14 +810,14 @@ class ReactExoplayerView extends FrameLayout implements
                  .setLoadErrorHandlingPolicy(
                         config.buildLoadErrorHandlingPolicy(minLoadRetryCount)
                 ).createMediaSource(mediaItem);
-            case C.TYPE_HLS:
+            case CONTENT_TYPE_HLS:
                 return new HlsMediaSource.Factory(
                         mediaDataSourceFactory
                 ).setDrmSessionManagerProvider(drmProvider)
                  .setLoadErrorHandlingPolicy(
                         config.buildLoadErrorHandlingPolicy(minLoadRetryCount)
                 ).createMediaSource(mediaItem);
-            case C.TYPE_OTHER:
+            case CONTENT_TYPE_OTHER:
                 return new ProgressiveMediaSource.Factory(
                         mediaDataSourceFactory
                 ).setDrmSessionManagerProvider(drmProvider)
@@ -810,12 +865,19 @@ class ReactExoplayerView extends FrameLayout implements
 
     private void releasePlayer() {
         if (player != null) {
+            if (adsLoader != null) {
+                adsLoader.setPlayer(null);
+            }
             updateResumePosition();
             player.release();
             player.removeListener(this);
             trackSelector = null;
             player = null;
         }
+        if (adsLoader != null) {
+            adsLoader.release();
+        }
+        adsLoader = null;
         progressHandler.removeMessages(SHOW_PROGRESS);
         themedReactContext.removeLifecycleEventListener(this);
         audioBecomingNoisyReceiver.removeListener();
@@ -1007,6 +1069,7 @@ class ReactExoplayerView extends FrameLayout implements
                 text += "ready";
                 eventEmitter.ready();
                 onBuffering(false);
+                clearProgressMessageHandler(); // ensure there is no other message
                 startProgressHandler();
                 videoLoaded();
                 if (selectTrackWhenReady && isUsingContentResolution) {
@@ -1046,7 +1109,7 @@ class ReactExoplayerView extends FrameLayout implements
     }
 
     private void videoLoaded() {
-        if (loadVideoStarted) {
+        if (!player.isPlayingAd() && loadVideoStarted) {
             loadVideoStarted = false;
             if (audioTrackType != null) {
                 setSelectedAudioTrack(audioTrackType, audioTrackValue);
@@ -1065,25 +1128,42 @@ class ReactExoplayerView extends FrameLayout implements
             // Properties that must be accessed on the main thread
             long duration = player.getDuration();
             long currentPosition = player.getCurrentPosition();
-            WritableArray audioTrackInfo = getAudioTrackInfo();
-            WritableArray textTrackInfo = getTextTrackInfo();
-            int trackRendererIndex = getTrackRendererIndex(C.TRACK_TYPE_VIDEO);
+            ArrayList<Track> audioTracks = getAudioTrackInfo();
+            ArrayList<Track> textTracks  = getTextTrackInfo();
 
-            ExecutorService es = Executors.newSingleThreadExecutor();
-            es.execute(new Runnable() {
-                @Override
-                public void run() {
-                    // To prevent ANRs caused by getVideoTrackInfo we run this on a different thread and notify the player only when we're done
-                    eventEmitter.load(duration, currentPosition, width, height,
-                        audioTrackInfo, textTrackInfo, getVideoTrackInfo(trackRendererIndex), trackId);
-                }
-            });
+            if (this.contentStartTime != -1L) {
+                ExecutorService es = Executors.newSingleThreadExecutor();
+                es.execute(new Runnable() {
+                    @Override
+                    public void run() {
+                        // To prevent ANRs caused by getVideoTrackInfo we run this on a different thread and notify the player only when we're done
+                        ArrayList<VideoTrack> videoTracks = getVideoTrackInfoFromManifest();
+                        if (videoTracks != null) {
+                            isUsingContentResolution = true;
+                        }
+                        eventEmitter.load(duration, currentPosition, width, height,
+                                audioTracks, textTracks, videoTracks, trackId );
+
+                    }
+                });
+                return;
+            }
+
+            ArrayList<VideoTrack> videoTracks = getVideoTrackInfo();
+
+            eventEmitter.load(duration, currentPosition, width, height,
+                    audioTracks, textTracks, videoTracks, trackId);
         }
     }
 
-    private WritableArray getAudioTrackInfo() {
-        WritableArray audioTracks = Arguments.createArray();
+    private static boolean isTrackSelected(TrackSelection selection, TrackGroup group,
+                                           int trackIndex){
+        return selection != null && selection.getTrackGroup() == group
+                && selection.indexOf( trackIndex ) != C.INDEX_UNSET;
+    }
 
+    private ArrayList<Track> getAudioTrackInfo() {
+        ArrayList<Track> audioTracks = new ArrayList<>();
         if (trackSelector == null) {
             // Likely player is unmounting so no audio tracks are available anymore
             return audioTracks;
@@ -1094,78 +1174,76 @@ class ReactExoplayerView extends FrameLayout implements
         if (info == null || index == C.INDEX_UNSET) {
             return audioTracks;
         }
-
         TrackGroupArray groups = info.getTrackGroups(index);
+        TrackSelectionArray selectionArray = player.getCurrentTrackSelections();
+        TrackSelection selection = selectionArray.get( C.TRACK_TYPE_AUDIO );
+
         for (int i = 0; i < groups.length; ++i) {
-            Format format = groups.get(i).getFormat(0);
-            WritableMap audioTrack = Arguments.createMap();
-            audioTrack.putInt("index", i);
-            audioTrack.putString("title", format.id != null ? format.id : "");
-            audioTrack.putString("type", format.sampleMimeType);
-            audioTrack.putString("language", format.language != null ? format.language : "");
-            audioTrack.putString("bitrate", format.bitrate == Format.NO_VALUE ? ""
-                                    : String.format(Locale.US, "%.2fMbps", format.bitrate / 1000000f));
-            audioTracks.pushMap(audioTrack);
+            TrackGroup group = groups.get(i);
+            Format format = group.getFormat(0);
+            Track audioTrack = new Track();
+            audioTrack.m_index = i;
+            audioTrack.m_title = format.id != null ? format.id : "";
+            audioTrack.m_mimeType = format.sampleMimeType;
+            audioTrack.m_language = format.language != null ? format.language : "";
+            audioTrack.m_bitrate = format.bitrate == Format.NO_VALUE ? 0 : format.bitrate;
+            audioTrack.m_isSelected = isTrackSelected(selection, group, 0 );
+            audioTracks.add(audioTrack);
         }
         return audioTracks;
     }
-    private WritableArray getVideoTrackInfo(int trackRendererIndex) {
 
-        if (this.contentStartTime != -1L) {
-            WritableArray contentVideoTracks = this.getVideoTrackInfoFromManifest();
-            if (contentVideoTracks != null) {
-                isUsingContentResolution = true;
-                return contentVideoTracks;
-            }
+    private ArrayList<VideoTrack> getVideoTrackInfo() {
+        ArrayList<VideoTrack> videoTracks = new ArrayList<>();
+        if (trackSelector == null) {
+            // Likely player is unmounting so no audio tracks are available anymore
+            return videoTracks;
         }
-
-        WritableArray videoTracks = Arguments.createArray();
-
         MappingTrackSelector.MappedTrackInfo info = trackSelector.getCurrentMappedTrackInfo();
-
-        if (info == null || trackRendererIndex == C.INDEX_UNSET) {
+        int index = getTrackRendererIndex(C.TRACK_TYPE_VIDEO);
+        if (info == null || index == C.INDEX_UNSET) {
             return videoTracks;
         }
 
-        TrackGroupArray groups = info.getTrackGroups(trackRendererIndex);
+        TrackGroupArray groups = info.getTrackGroups(index);
         for (int i = 0; i < groups.length; ++i) {
             TrackGroup group = groups.get(i);
 
             for (int trackIndex = 0; trackIndex < group.length; trackIndex++) {
                 Format format = group.getFormat(trackIndex);
                 if (isFormatSupported(format)) {
-                    WritableMap videoTrack = Arguments.createMap();
-                    videoTrack.putInt("width", format.width == Format.NO_VALUE ? 0 : format.width);
-                    videoTrack.putInt("height",format.height == Format.NO_VALUE ? 0 : format.height);
-                    videoTrack.putInt("bitrate", format.bitrate == Format.NO_VALUE ? 0 : format.bitrate);
-                    videoTrack.putString("codecs", format.codecs != null ? format.codecs : "");
-                    videoTrack.putString("trackId", format.id == null ? String.valueOf(trackIndex) : format.id);
-                    videoTracks.pushMap(videoTrack);
+                    VideoTrack videoTrack = new VideoTrack();
+                    videoTrack.m_width = format.width == Format.NO_VALUE ? 0 : format.width;
+                    videoTrack.m_height = format.height == Format.NO_VALUE ? 0 : format.height;
+                    videoTrack.m_bitrate = format.bitrate == Format.NO_VALUE ? 0 : format.bitrate;
+                    videoTrack.m_codecs = format.codecs != null ? format.codecs : "";
+                    videoTrack.m_trackId = format.id == null ? String.valueOf(trackIndex) : format.id;
+                    videoTracks.add(videoTrack);
                 }
             }
         }
-
         return videoTracks;
     }
 
-    private WritableArray getVideoTrackInfoFromManifest() {
+    private ArrayList<VideoTrack> getVideoTrackInfoFromManifest() {
         return this.getVideoTrackInfoFromManifest(0);
     }
 
     // We need retry count to in case where minefest request fails from poor network conditions
-    private WritableArray getVideoTrackInfoFromManifest(int retryCount) {
+    @WorkerThread
+    private ArrayList<VideoTrack> getVideoTrackInfoFromManifest(int retryCount) {
         ExecutorService es = Executors.newSingleThreadExecutor();
         final DataSource dataSource = this.mediaDataSourceFactory.createDataSource();
         final Uri sourceUri = this.srcUri;
         final long startTime = this.contentStartTime * 1000 - 100; // s -> ms with 100ms offset
 
-        Future<WritableArray> result = es.submit(new Callable<WritableArray>() {
+        Future<ArrayList<VideoTrack>> result = es.submit(new Callable<ArrayList<VideoTrack>>() {
             DataSource ds = dataSource;
             Uri uri = sourceUri;
             long startTimeUs = startTime * 1000; // ms -> us
 
-            public WritableArray call() throws Exception {
-                WritableArray videoTracks = Arguments.createArray();
+            public ArrayList<VideoTrack> call() throws Exception {
+                ArrayList<VideoTrack> videoTracks = new ArrayList<>();
                 try  {
                     DashManifest manifest = DashUtil.loadManifest(this.ds, this.uri);
                     int periodCount = manifest.getPeriodCount();
@@ -1180,19 +1258,18 @@ class ReactExoplayerView extends FrameLayout implements
                             for (int representationIndex = 0; representationIndex < adaptation.representations.size(); representationIndex++) {
                                 Representation representation = adaptation.representations.get(representationIndex);
                                 Format format = representation.format;
-                                if (representation.presentationTimeOffsetUs <= startTimeUs) {
-                                    break;
-                                }
-                                hasFoundContentPeriod = true;
-                                WritableMap videoTrack = Arguments.createMap();
-                                videoTrack.putInt("width", format.width == Format.NO_VALUE ? 0 : format.width);
-                                videoTrack.putInt("height",format.height == Format.NO_VALUE ? 0 : format.height);
-                                videoTrack.putInt("bitrate", format.bitrate == Format.NO_VALUE ? 0 : format.bitrate);
-                                videoTrack.putString("codecs", format.codecs != null ? format.codecs : "");
-                                videoTrack.putString("trackId",
-                                        format.id == null ? String.valueOf(representationIndex) : format.id);
                                 if (isFormatSupported(format)) {
-                                    videoTracks.pushMap(videoTrack);
+                                    if (representation.presentationTimeOffsetUs <= startTimeUs) {
+                                        break;
+                                    }
+                                    hasFoundContentPeriod = true;
+                                    VideoTrack videoTrack = new VideoTrack();
+                                    videoTrack.m_width = format.width == Format.NO_VALUE ? 0 : format.width;
+                                    videoTrack.m_height = format.height == Format.NO_VALUE ? 0 : format.height;
+                                    videoTrack.m_bitrate = format.bitrate == Format.NO_VALUE ? 0 : format.bitrate;
+                                    videoTrack.m_codecs = format.codecs != null ? format.codecs : "";
+                                    videoTrack.m_trackId = format.id == null ? String.valueOf(representationIndex) : format.id;
+                                    videoTracks.add(videoTrack);
                                 }
                             }
                             if (hasFoundContentPeriod) {
@@ -1206,7 +1283,7 @@ class ReactExoplayerView extends FrameLayout implements
         });
 
         try {
-            WritableArray results = result.get(3000, TimeUnit.MILLISECONDS);
+            ArrayList<VideoTrack> results = result.get(3000, TimeUnit.MILLISECONDS);
             if (results == null && retryCount < 1) {
                 return this.getVideoTrackInfoFromManifest(++retryCount);
             }
@@ -1217,24 +1294,31 @@ class ReactExoplayerView extends FrameLayout implements
         return null;
     }
 
-    private WritableArray getTextTrackInfo() {
-        WritableArray textTracks = Arguments.createArray();
-
+    private ArrayList<Track> getTextTrackInfo() {
+        ArrayList<Track> textTracks = new ArrayList<>();
+        if (trackSelector == null) {
+            return textTracks;
+        }
         MappingTrackSelector.MappedTrackInfo info = trackSelector.getCurrentMappedTrackInfo();
         int index = getTrackRendererIndex(C.TRACK_TYPE_TEXT);
         if (info == null || index == C.INDEX_UNSET) {
             return textTracks;
         }
-
+        TrackSelectionArray selectionArray = player.getCurrentTrackSelections();
+        TrackSelection selection = selectionArray.get( C.TRACK_TYPE_VIDEO );
         TrackGroupArray groups = info.getTrackGroups(index);
+
         for (int i = 0; i < groups.length; ++i) {
-             Format format = groups.get(i).getFormat(0);
-             WritableMap textTrack = Arguments.createMap();
-             textTrack.putInt("index", i);
-             textTrack.putString("title", format.id != null ? format.id : "");
-             textTrack.putString("type", format.sampleMimeType);
-             textTrack.putString("language", format.language != null ? format.language : "");
-             textTracks.pushMap(textTrack);
+            TrackGroup group = groups.get(i);
+            Format format = group.getFormat(0);
+
+            Track textTrack = new Track();
+            textTrack.m_index = i;
+            textTrack.m_title = format.id != null ? format.id : "";
+            textTrack.m_mimeType = format.sampleMimeType;
+            textTrack.m_language = format.language != null ? format.language : "";
+            textTrack.m_isSelected = isTrackSelected(selection, group, 0 );
+            textTracks.add(textTrack);
         }
         return textTracks;
     }
@@ -1303,7 +1387,9 @@ class ReactExoplayerView extends FrameLayout implements
 
     @Override
     public void onTracksChanged(Tracks tracks) {
-        // Do nothing.
+        eventEmitter.textTracks(getTextTrackInfo());
+        eventEmitter.audioTracks(getAudioTrackInfo());
+        eventEmitter.videoTracks(getVideoTrackInfo());
     }
 
     @Override
@@ -1416,6 +1502,10 @@ class ReactExoplayerView extends FrameLayout implements
         mReportBandwidth = reportBandwidth;
     }
 
+    public void setAdTagUrl(final Uri uri) {
+        adTagUrl = uri;
+    }
+
     public void setRawSrc(final Uri uri, final String extension) {
         if (uri != null) {
             boolean isSourceEqual = uri.equals(srcUri);
@@ -1463,6 +1553,14 @@ class ReactExoplayerView extends FrameLayout implements
         this.preventsDisplaySleepDuringVideoPlayback = preventsDisplaySleepDuringVideoPlayback;
     }
 
+    public void disableTrack(int rendererIndex) {
+        DefaultTrackSelector.Parameters disableParameters = trackSelector.getParameters()
+                .buildUpon()
+                .setRendererDisabled(rendererIndex, true)
+                .build();
+        trackSelector.setParameters(disableParameters);
+    }
+
     public void setSelectedTrack(int trackType, String type, Dynamic value) {
         if (player == null) return;
         int rendererIndex = getTrackRendererIndex(trackType);
@@ -1483,13 +1581,8 @@ class ReactExoplayerView extends FrameLayout implements
             type = "default";
         }
 
-        DefaultTrackSelector.Parameters disableParameters = trackSelector.getParameters()
-                .buildUpon()
-                .setRendererDisabled(rendererIndex, true)
-                .build();
-
         if (type.equals("disabled")) {
-            trackSelector.setParameters(disableParameters);
+            disableTrack(rendererIndex);
             return;
         } else if (type.equals("language")) {
             for (int i = 0; i < groups.length; ++i) {
@@ -1607,7 +1700,7 @@ class ReactExoplayerView extends FrameLayout implements
         }
 
         if (groupIndex == C.INDEX_UNSET) {
-            trackSelector.setParameters(disableParameters);
+            disableTrack(rendererIndex);
             return;
         }
 
@@ -1616,6 +1709,7 @@ class ReactExoplayerView extends FrameLayout implements
         DefaultTrackSelector.Parameters selectionParameters = trackSelector.getParameters()
             .buildUpon()
             .setRendererDisabled(rendererIndex, false)
+            .clearOverridesOfType(selectionOverride.getType())
             .addOverride(selectionOverride)
             .build();
         trackSelector.setParameters(selectionParameters);
@@ -1767,6 +1861,22 @@ class ReactExoplayerView extends FrameLayout implements
         this.disableBuffering = disableBuffering;
     }
 
+    private void updateFullScreenButtonVisbility() {
+        if (playerControlView != null) {
+            final ImageButton fullScreenButton = playerControlView.findViewById(R.id.exo_fullscreen);
+            if (controls) {
+                //Handling the fullScreenButton click event
+                if (isFullscreen && fullScreenPlayerView != null && !fullScreenPlayerView.isShowing()) {
+                    fullScreenButton.setVisibility(GONE);
+                } else {
+                    fullScreenButton.setVisibility(VISIBLE);
+                }
+            } else {
+                fullScreenButton.setVisibility(GONE);
+            }
+        }
+    }
+
     public void setDisableDisconnectError(boolean disableDisconnectError) {
         this.disableDisconnectError = disableDisconnectError;
     }
@@ -1782,15 +1892,6 @@ class ReactExoplayerView extends FrameLayout implements
             return;
         }
 
-        if (fullScreenPlayerView == null) {
-            fullScreenPlayerView = new FullScreenPlayerView(getContext(), exoPlayerView, playerControlView, new OnBackPressedCallback(true) {
-                @Override
-                public void handleOnBackPressed() {
-                    setFullscreen(false);
-                }
-            });
-        }
-
         Window window = activity.getWindow();
         View decorView = window.getDecorView();
         int uiOptions;
@@ -1804,7 +1905,7 @@ class ReactExoplayerView extends FrameLayout implements
                         | SYSTEM_UI_FLAG_FULLSCREEN;
             }
             eventEmitter.fullscreenWillPresent();
-            if (controls) {
+            if (controls && fullScreenPlayerView != null) {
                 fullScreenPlayerView.show();
             }
             post(() -> {
@@ -1814,7 +1915,7 @@ class ReactExoplayerView extends FrameLayout implements
         } else {
             uiOptions = View.SYSTEM_UI_FLAG_VISIBLE;
             eventEmitter.fullscreenWillDismiss();
-            if (controls) {
+            if (controls && fullScreenPlayerView != null) {
                 fullScreenPlayerView.dismiss();
                 reLayout(exoPlayerView);
             }
@@ -1823,6 +1924,8 @@ class ReactExoplayerView extends FrameLayout implements
                 eventEmitter.fullscreenDidDismiss();
             });
         }
+        // need to be done at the end to avoid hiding fullscreen control button when fullScreenPlayerView is shown
+        updateFullScreenButtonVisbility();
     }
 
     public void setUseTextureView(boolean useTextureView) {
@@ -1891,9 +1994,9 @@ class ReactExoplayerView extends FrameLayout implements
      */
     public void setControls(boolean controls) {
         this.controls = controls;
-        if (player == null || exoPlayerView == null) return;
         if (controls) {
             addPlayerControl();
+            updateFullScreenButtonVisbility();
         } else {
             int indexOfPC = indexOfChild(playerControlView);
             if (indexOfPC != -1) {
@@ -1904,5 +2007,10 @@ class ReactExoplayerView extends FrameLayout implements
 
     public void setSubtitleStyle(SubtitleStyle style) {
         exoPlayerView.setSubtitleStyle(style);
+    }
+
+    @Override
+    public void onAdEvent(AdEvent adEvent) {
+        eventEmitter.receiveAdEvent(adEvent.getType().name());
     }
 }