+++ android/src/main/java/com/brentvatne/exoplayer/ReactExoplayerViewManager.java
@@ -31,6 +31,7 @@ public class ReactExoplayerViewManager extends ViewGroupManager<ReactExoplayerVi
 
     private static final String PROP_SRC = "src";
     private static final String PROP_SRC_URI = "uri";
+    private static final String PROP_AD_TAG_URL = "adTagUrl";
     private static final String PROP_SRC_TYPE = "type";
     private static final String PROP_DRM = "drm";
     private static final String PROP_DRM_TYPE = "type";
@@ -189,6 +190,18 @@ public class ReactExoplayerViewManager extends ViewGroupManager<ReactExoplayerVi
         }
     }
 
+    @ReactProp(name = PROP_AD_TAG_URL)
+    public void setAdTagUrl(final ReactExoplayerView videoView, final String uriString) {
+        if (TextUtils.isEmpty(uriString)) {
+            return;
+        }
+
+        Uri adTagUrl = Uri.parse(uriString);
+
+        videoView.setAdTagUrl(adTagUrl);
+    }
+
+
     @ReactProp(name = PROP_RESIZE_MODE)
     public void setResizeMode(final ReactExoplayerView videoView, final String resizeModeOrdinalString) {
         videoView.setResizeModeModifier(convertToIntDef(resizeModeOrdinalString));
@@ -317,7 +330,7 @@ public class ReactExoplayerViewManager extends ViewGroupManager<ReactExoplayerVi
         videoView.setBackBufferDurationMs(backBufferDurationMs);
     }
 
-    @ReactProp(name = PROP_CONTENT_START_TIME, defaultInt = 0)
+    @ReactProp(name = PROP_CONTENT_START_TIME, defaultInt = -1)
     public void setContentStartTime(final ReactExoplayerView videoView, final int contentStartTime) {
         videoView.setContentStartTime(contentStartTime);
     }